backend/training/requirements-train.txt
tensorflow-cpu==2.16.1
tensorflow-datasets==4.9.6
numpy==1.26.4
